refactor: Movie -> movie 이름 변경
패키지명은 소문자를 사용하는 관습에 따라 기존의 디렉토리명을 Movie -> movie 변경했습니다.
