+++ movie-book/src/main/java/hello/moviebook/Movie/domain/Genre.java
@@ -1,4 +1,4 @@
-package hello.moviebook.Movie;
+package hello.moviebook.movie.domain;
 
 import jakarta.persistence.*;
