+++ movie-book/src/main/java/hello/moviebook/Movie/domain/UserGenre.java
@@ -1,6 +1,6 @@
-package hello.moviebook.UserGenre;
+package hello.moviebook.movie.domain;
 
-import hello.moviebook.User.User;
+import hello.moviebook.user.domain.User;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
