+++ movie-book/src/main/java/hello/moviebook/Movie/repository/UserGenreRepository.java
@@ -1,6 +1,7 @@
-package hello.moviebook.UserGenre;
+package hello.moviebook.movie.repository;
 
-import hello.moviebook.User.User;
+import hello.moviebook.user.domain.User;
+import hello.moviebook.movie.domain.UserGenre;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
