+++ movie-book/src/main/java/hello/moviebook/Movie/repository/UserMovieRepository.java
@@ -1,5 +1,6 @@
-package hello.moviebook.UserMovie;
+package hello.moviebook.movie.repository;
 
+import hello.moviebook.movie.domain.UserMovie;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
